Editing memories, Constants class, removed Dummy content, FileUtils
Memories can now be edited (via a floating action button in the detail view of the memory). Editing uses the same activity as adding a memory but the EditTexts are pre-populated with the current values for the memory. If a memory is edited, "saving" the edit (by creating the memory) actually creates a new memory. This will still need to be updated (and change the button text to be specific for editing).
The Constants class will be used to store some values that are needed by multiple classes for easier access.
FileUtils has a method to update the memories.csv file (for instance, if the delimiter is changed) so that deleting the entire file and creating a new one is no longer needed. It also has a method (not yet used or tested) which essentially deletes lines from a file if their content equals a given string (this will be used for deleting memories).