Memories can be deleted from the detail view.
Deletion works well now (TODO: a confirmation should be added later and a setting to turn it off/on).
There are one or two known errors at present: adding a memory didn't lead to it being shown right away; going to a detail view and then back after adding a new memory caused a fatal error. This error is either gone or now masked by a new error: trying to create a new memory at all throws a fatal error.