External memory usage, exporting memories file.
Added support for storing memories file in external storage and for exporting that file (sending the .csv by email).
This broke the implementation that used an internal file; it should be fixed sometime but is a low priority now.

Also fixed the memory detail screen so that when it is rotated, the title is still set appropriately.