Memories are shown sorted by date, search works, bug fixes and cleanup.
A SortedList is now used (with a comparator by date) so that memories are shown in date order in the RecyclerView.

Searching is implemented, albeit somewhat naively. Searching for multiple tags requires them to be in the same order in the memory as they are in your search query. Besides this, it works rather well (the filter used when searching shows all memories whose String representation contains the query; thus the need for the ordering of the query to match that inside of the Memory).

All FileUtils was made a bit nicer; all creating/deleting/editing of Memorys were moved to functions within Memory which update the lists appropriately and call the relevant FileUtils methods. Editing a Memory no longer does a delete row and a create row; instead, the previous row is edited. This behavior is more robust; the file now contains the memories in exactly the order they were created in.

A bit more documentation was added (though this area is still somewhat lacking) and a couple of bugs were fixed by the above changes (e.g. the one with creating a new memory, editing it, then leaving the app).