Adding memories works!
A ListView.Builder is used instead of just the ListView so that when new MemoryWidgets are added, the display updates accordingly.
The Memory class was renamed to MemoryWidget and a new Memory class (which isn't a widget) was created. Decoupling this makes it so we can change the attributes of a Memory, which we do while constructing it in AddMemory (see below). Once a Memory is finished, it can be wrapped with MemoryWidget to add on the display information.
The AddMemory page creates an empty Memory (if editing a memory, the values from the memory being edited are used instead) which is updated whenever the TextFields (date, tags, text) are updated. When Save is pressed, a MemoryWidget is created from this Memory and added to the list.
MemoryListState now defines methods for adding new memories and updating (i.e. overwriting) existing ones. Defining these methods in a State lets us call setState whenever the list of memories is updated so that the UI will be refreshed.