Using Flux to store the memories.
Especially when implementing the edit page, I had to pass a reference to the MemoryListState through several classes. This approach seemed suboptimal, and at some point (despite getting no errors) the MemoryList page didn't show any of the memories. To alleviate issues with needing access to the memories in multiple locations, memory_store now defines a Store which manages the list of Memory's. It defines the actions that can be used to add or update a Memory (the Memory class itself was also moved to memory_store).

MemoryListState listens to MemoryStore and updates its view accordingly. The AddMemory page is working now with Flux, but the EditMemory page isn't complete yet.