Edit page is functional.
The Edit page now correctly sets the initial values of the text fields to be the current values for the memory being edited. Saving an edit navigates back to the MemoryList page, as it did in the old implementation (going back to the MemoryDetail screen might be better, but we'd need to refresh the UI there to show the edit to that memory).