Edit FAB moved to AppBar, index update fix.
Editing memories isn't that common, so I moved the FAB to be an IconButton in the AppBar instead; this makes it less obtrusive.
There was also a bug (a race condition?) because a top-level MemoryStore was instantiated directly in main.dart in addition to the MemoryStore that MemoryListState listened to. This seems to have created two instances which both handled all actions. Depending on the timing of their action handling, their state could get out of sync. AddEditMemoryState now also uses the StoreWatcherMixin and listens to the MemoryStore properly by its token, so there's only one store and actions are taken exactly once.