Moved (de)serialization of memories into Memory class.
No functionality has been changed, but the code is a little nicer this way; MemoryStore doesn't know or care exactly how to serialize a Memory. Instead, a Memory knows how to convert itself into a String for serialization as well as how to convert such a String back into a Memory. This should make this section more maintainable during changes to serialization.