Importing memories files, minor UI improvements.
There doesn't appear to be a nice way to open a file using core Flutter, so I had to introduce a dependency on somebody's solution for this that doesn't seem much used yet; it seems to work well, still.
All lines from an imported file that can be deserialized into memories are added to the current memories list (and thus also file).
The main page now only shows the first 50 characters of the text of each memory (and 20 of the tags) so that there aren't overflow issues.

Problems:
 - scrolling on the main page is bad; in the old app, it felt very smooth, but now it seems the frame rate is too low, and it's jerky. This problem probably existed before, but I never tested with enough memories to be able to scroll.