Add `TextOverflow` to memory tags in list.
This lets us, as done before with the memory prefixes, remove the explicit length restriction and just use an ellipsis when the tags are too long to show. We use `Expanded` widgets with the `flex` attribute to control how the screen width is shared between the date/tags column and the memory prefix.