Add search in app bar.
Clicking the search icon will replace the app bar title with a `TextField` where you can enter the search query and replace the app bar actions with a close button to exit searching. The `MemoryStore` now has a function to get a list of memories which match some search query.

I think there may be one issue with this implementation, though. If a new `Memory` is added to the store, I'm not sure if that will be reflected right away on the main page; this needs testing.