Scroll long memories, `setState` after add/edit.
The tags are also now split on ", " instead of ","; this should remove extra whitespace.
The text on the `MemoryDetail` screen scrolls so that long memories are viewable.
The memory list is `RouteAware` so that it can call `setState` when you navigate back to it. This means that adding/editing a memory will be shown right away. There's probably a better way to do this using the store itself, especially because the current method requires a delay for a hard-coded amount of time that's only been verified to work on my phone.